feat: free discuss action btn

+++ audio_recorder.py
@@ -47,13 +47,15 @@ class AudioRecorder:
             while self.recording:
                 data = self.stream.read(self.chunk)
                 self.output_stream.write(data)
-                self.frames.append(data)
+                if(self.record_user_obj['is_free_discuss']=="false"):
+                   self.frames.append(data)
             
             self.stream.stop_stream()
             self.stream.close()
             self.output_stream.stop_stream()
             self.output_stream.close()
-            self.save_wave()
+            if(self.record_user_obj['is_free_discuss']=="false"):
+             self.save_wave()
         
         self.record_thread = threading.Thread(target=record)
         self.record_thread.start()
@@ -80,7 +82,7 @@ class AudioRecorder:
 
     def terminate(self):
         self.audio.terminate()   
-        if(os.path.exists(self.output_audio_path)):
+        if(os.path.exists(self.output_audio_path) and self.record_user_obj['is_free_discuss']=="false"):
             file_upload_service.file_upload_to_server(self.record_user_obj['usercode'],self.record_user_obj)   
             file_upload_service.delete_file_after_upload(self.output_audio_path)
             file_upload_service.delete_file_after_upload(self.record_user_obj['usercode'])
